Fix zero3 tracing issues
